requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.4
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via pip-tools
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==3.2
    # via
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
first==2.0.2
    # via pip-tools
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
packaging==20.9
    # via drf-yasg
pip-tools==1.8.0
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via django
requests==2.25.1
    # via coreapi
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.4
    # via drf-yasg
six==1.15.0
    # via pip-tools
sqlparse==0.4.1
    # via django
typing-extensions==3.7.4.3
    # via asgiref
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.4
    # via requests
